refactor: Mock 아티스트, 학교 생성 로직 분리
