+++ backend/src/main/java/com/festago/mock/domain/MockDataInitializer.java
@@ -1,0 +1,8 @@
+package com.festago.mock.domain;
+
+public interface MockDataInitializer {
+
+    boolean canInitialize();
+
+    void initialize();
+}
